separate nnue related files
