--- src/Raphael/nnue.cpp
@@ -1,669 +1,0 @@
-#include <Raphael/nnue.h>
-
-#define INCBIN_PREFIX g_
-#define INCBIN_STYLE INCBIN_STYLE_SNAKE
-#include <thirdparty/incbin.h>
-
-#include <fstream>
-#include <stdexcept>
-
-using namespace raphael;
-using std::copy;
-using std::max;
-using std::min;
-using std::ofstream;
-using std::popcount;
-using std::runtime_error;
-using std::vector;
-
-#define STRINGIFY(x) #x
-#define TOSTRING(x) STRINGIFY(x)
-
-INCBIN(unsigned char, netfile, TOSTRING(NETWORK_FILE));
-
-
-
-i32 Nnue::PSQFeature::index(chess::Color perspective, bool mirror) const {
-    const auto sq = (mirror) ? square.mirrored() : square;
-    const auto pc = (piece.type() == chess::PieceType::KING) ? chess::Piece::WHITEKING
-                                                             : piece.relative(perspective);
-    return 64 * pc + sq.relative(perspective);
-}
-
-
-
-void Nnue::NnueFinnyEntry::initialize(const i16 biases[L1_SIZE]) {
-    copy(biases, biases + L1_SIZE, values);
-}
-
-void Nnue::NnueFinnyEntry::sync(
-    const i16 weights[N_INPUTS][L1_SIZE],
-    const chess::Board& board,
-    chess::Color perspective,
-    bool mirror
-) {
-    i32 adds[32];
-    i32 subs[32];
-    i32 n_adds = 0;
-    i32 n_subs = 0;
-
-    // compute diff from finny_entry
-    for (chess::PieceType pt = chess::PieceType::PAWN; pt <= chess::PieceType::KING; ++pt) {
-        for (const auto color : {chess::Color::WHITE, chess::Color::BLACK}) {
-            const auto old_occ = occ(pt, color);
-            const auto new_occ = board.occ(pt, color);
-
-            auto adds_occ = new_occ & ~old_occ;
-            while (adds_occ) {
-                const auto sq = chess::Square(adds_occ.poplsb());
-                const auto piece = chess::Piece(pt, color);
-
-                assert(n_adds < 32);
-                adds[n_adds++] = PSQFeature(piece, sq).index(perspective, mirror);
-            }
-
-            auto subs_occ = old_occ & ~new_occ;
-            while (subs_occ) {
-                const auto sq = chess::Square(subs_occ.poplsb());
-                const auto piece = chess::Piece(pt, color);
-
-                assert(n_subs < 32);
-                subs[n_subs++] = PSQFeature(piece, sq).index(perspective, mirror);
-            }
-        }
-    }
-
-    // update occupancy bitboards
-    for (chess::PieceType pt = chess::PieceType::PAWN; pt <= chess::PieceType::KING; ++pt)
-        pieces_[pt] = board.occ(pt);
-    for (const auto color : {chess::Color::WHITE, chess::Color::BLACK})
-        occ_[color] = board.occ(color);
-
-#ifdef USE_SIMD
-    constexpr i32 regw = ALIGNMENT / sizeof(i16);
-    constexpr i32 n_chunks = L1_SIZE / regw;
-    static_assert(L1_SIZE % regw == 0);
-    static_assert(n_chunks % 8 == 0);
-    VecI16 accs[8];
-
-    for (i32 i = 0; i < n_chunks; i += 8) {
-        #pragma GCC unroll 32
-        for (i32 r = 0; r < 8; r++) accs[r] = load_i16(&values[(i + r) * regw]);
-
-        // add features
-        for (i32 f = 0; f < n_adds; f++) {
-            const auto fidx = adds[f];
-
-            #pragma GCC unroll 32
-            for (i32 r = 0; r < 8; r++)
-                accs[r] = add_i16(accs[r], load_i16(&weights[fidx][(i + r) * regw]));
-        }
-
-        // rem features
-        for (i32 f = 0; f < n_subs; f++) {
-            const auto fidx = subs[f];
-
-            #pragma GCC unroll 32
-            for (i32 r = 0; r < 8; r++)
-                accs[r] = sub_i16(accs[r], load_i16(&weights[fidx][(i + r) * regw]));
-        }
-
-        #pragma GCC unroll 32
-        for (i32 r = 0; r < 8; r++) store_i16(&values[(i + r) * regw], accs[r]);
-    }
-#else
-    for (i32 f = 0; f < n_adds; f++)
-        for (i32 i = 0; i < L1_SIZE; i++) values[i] += weights[adds[f]][i];
-    for (i32 f = 0; f < n_subs; f++)
-        for (i32 i = 0; i < L1_SIZE; i++) values[i] -= weights[subs[f]][i];
-#endif
-}
-
-chess::BitBoard Nnue::NnueFinnyEntry::occ(chess::PieceType pt, chess::Color color) const {
-    return pieces_[pt] & occ_[color];
-}
-
-
-
-Nnue::NnueAccumulator::PSQState Nnue::NnueAccumulator::get_psq_state(
-    chess::Color perspective
-) const {
-    return psq_state[perspective];
-}
-
-void Nnue::NnueAccumulator::set_psq_state(chess::Color perspective, PSQState state) {
-    psq_state[perspective] = state;
-}
-
-void Nnue::NnueAccumulator::add_piece(chess::Piece piece, chess::Square square) {
-    psq_adds.push({.piece = piece, .square = square});
-}
-
-void Nnue::NnueAccumulator::rem_piece(chess::Piece piece, chess::Square square) {
-    psq_subs.push({.piece = piece, .square = square});
-}
-
-void Nnue::NnueAccumulator::prepare_updates() {
-    // reset psq updates and mark as dirty (as we're going to update them immediately after)
-    psq_adds.clear();
-    psq_subs.clear();
-    set_psq_state(chess::Color::WHITE, PSQState::DIRTY);
-    set_psq_state(chess::Color::BLACK, PSQState::DIRTY);
-}
-
-void Nnue::NnueAccumulator::apply_updates(
-    const NnueAccumulator& old_acc,
-    const i16 weights[N_INPUTS][L1_SIZE],
-    chess::Color perspective,
-    bool mirror
-) {
-    assert(get_psq_state(perspective) == PSQState::DIRTY);
-    assert(old_acc.get_psq_state(perspective) == PSQState::CLEAN);
-    assert(psq_adds.size() >= 1);
-    assert(psq_subs.size() >= 1);
-
-    const i32 add1 = psq_adds[0].index(perspective, mirror);
-    const i32 add2 = (psq_adds.size() > 1) ? psq_adds[1].index(perspective, mirror) : 0;
-    const i32 sub1 = psq_subs[0].index(perspective, mirror);
-    const i32 sub2 = (psq_subs.size() > 1) ? psq_subs[1].index(perspective, mirror) : 0;
-
-#ifdef USE_SIMD
-    constexpr i32 regw = ALIGNMENT / sizeof(i16);
-    constexpr i32 n_chunks = L1_SIZE / regw;
-    static_assert(L1_SIZE % regw == 0);
-    static_assert(n_chunks % 8 == 0);
-    VecI16 accs[8];
-
-    for (i32 i = 0; i < n_chunks; i += 8) {
-        #pragma GCC unroll 32
-        for (i32 r = 0; r < 8; r++)
-            accs[r] = load_i16(&old_acc.values[perspective][(i + r) * regw]);
-
-        #pragma GCC unroll 32
-        for (i32 r = 0; r < 8; r++)
-            accs[r] = sub_i16(accs[r], load_i16(&weights[sub1][(i + r) * regw]));
-
-        if (psq_subs.size() > 1)
-            #pragma GCC unroll 32
-            for (i32 r = 0; r < 8; r++)
-                accs[r] = sub_i16(accs[r], load_i16(&weights[sub2][(i + r) * regw]));
-
-        #pragma GCC unroll 32
-        for (i32 r = 0; r < 8; r++)
-            accs[r] = add_i16(accs[r], load_i16(&weights[add1][(i + r) * regw]));
-
-        if (psq_adds.size() > 1)
-            #pragma GCC unroll 32
-            for (i32 r = 0; r < 8; r++)
-                accs[r] = add_i16(accs[r], load_i16(&weights[add2][(i + r) * regw]));
-
-        #pragma GCC unroll 32
-        for (i32 r = 0; r < 8; r++) store_i16(&values[perspective][(i + r) * regw], accs[r]);
-    }
-#else
-    for (i32 i = 0; i < L1_SIZE; i++) {
-        values[perspective][i] = old_acc.values[perspective][i];
-
-        values[perspective][i] -= weights[sub1][i];
-        if (psq_subs.size() > 1) values[perspective][i] -= weights[sub2][i];
-        values[perspective][i] += weights[add1][i];
-        if (psq_adds.size() > 1) values[perspective][i] += weights[add2][i];
-    }
-#endif
-
-    // mark as clean
-    set_psq_state(perspective, PSQState::CLEAN);
-}
-
-void Nnue::NnueAccumulator::refresh_from(
-    const NnueFinnyEntry& finny_entry, chess::Color perspective
-) {
-    copy(finny_entry.values, finny_entry.values + L1_SIZE, values[perspective]);
-    set_psq_state(perspective, PSQState::CLEAN);
-}
-
-#ifdef USE_SIMD
-void Nnue::SparseIterator::add_nonzeros(VecU8 l0_out0, VecU8 l0_out1) {
-    constexpr i32 regw32 = ALIGNMENT / sizeof(i32);
-    static_assert(regw32 % 8 == 0);
-
-    #ifdef __AVX512VBMI2__
-    static_assert(USE_SIMD == 512);
-    const auto mask = _mm512_kunpackw(nonzero_mask(l0_out1), nonzero_mask(l0_out0));
-
-    const auto idxs = _mm512_maskz_compress_epi16(mask, offset_);
-    _mm512_storeu_si512(&indices_[count_], idxs);
-    offset_ = add_i16(offset_, full_i16(32));
-    count_ += popcount(mask);
-
-    #else
-    u32 full_mask = (nonzero_mask(l0_out1) << regw32) | nonzero_mask(l0_out0);
-
-    for (i32 i = 0; i < regw32 / 4; i++) {
-        // get offset of up to 8 nonzeros at a time
-        const u8 mask = full_mask & 0xFF;
-        full_mask >>= 8;
-
-        const auto idxs = _mm_add_epi16(
-            offset_, _mm_load_si128(reinterpret_cast<const __m128i*>(&nonzero_idx[mask]))
-        );
-        _mm_storeu_si128(reinterpret_cast<__m128i*>(&indices_[count_]), idxs);
-        offset_ = _mm_add_epi16(offset_, _mm_set1_epi16(8));
-        count_ += popcount(mask);
-    }
-    #endif
-
-    assert(count_ <= L1_SIZE / 4);
-}
-
-i32 Nnue::SparseIterator::count() const { return count_; }
-
-i32 Nnue::SparseIterator::index(i32 nnz_id) const {
-    assert(nnz_id < count_);
-    return indices_[nnz_id];
-}
-#endif
-
-
-
-Nnue::Nnue(): params(load_network()), idx_(0) {
-    // set the finny table entries to the bias
-    for (const auto perspective : {chess::Color::WHITE, chess::Color::BLACK})
-        for (const auto mirror : {false, true})
-            for (i32 bucket = 0; bucket < N_INBUCKETS; bucket++)
-                finny_table[perspective][mirror][bucket].initialize(params->b0);
-}
-
-const Nnue::NnueParams* Nnue::load_network() {
-    constexpr usize padded_size = 64 * ((sizeof(NnueParams) + 63) / 64);
-    if (g_netfile_size != padded_size)
-        throw runtime_error("network file and architecture doesn't match");
-
-    if (reinterpret_cast<uintptr_t>(g_netfile_data) % alignof(NnueParams) != 0)
-        throw runtime_error("network file isn't aligned properly");
-
-    return reinterpret_cast<const NnueParams*>(g_netfile_data);
-}
-
-
-
-i32 Nnue::evaluate(const chess::Board& board) {
-    // apply lazy updates
-    lazy_update(board, chess::Color::WHITE);
-    lazy_update(board, chess::Color::BLACK);
-
-    // get address to accumulators
-    assert(accumulators[idx_].get_psq_state(board.stm()) == NnueAccumulator::PSQState::CLEAN);
-    assert(accumulators[idx_].get_psq_state(~board.stm()) == NnueAccumulator::PSQState::CLEAN);
-    const auto stm_acc = accumulators[idx_].values[board.stm()];
-    const auto ntm_acc = accumulators[idx_].values[~board.stm()];
-
-    alignas(ALIGNMENT) u8 l0_out[L1_SIZE];
-    SparseIterator sp;
-    activate_l0(stm_acc, l0_out, sp);
-    activate_l0(ntm_acc, l0_out + L1_SIZE / 2, sp);
-
-    constexpr i32 bucket_div = (32 + N_OUTBUCKETS - 1) / N_OUTBUCKETS;
-    const i32 bucket_idx = (board.occ().count() - 2) / bucket_div;
-
-    alignas(ALIGNMENT) i32 l1_out[L2_SIZE];
-    i64 eval;
-    forward_l1(l0_out, l1_out, sp, bucket_idx);
-    forward_l2l3(l1_out, eval, bucket_idx);
-
-    eval *= OUTPUT_SCALE;
-    eval /= (QC * QC * QC * QC);
-    return static_cast<i32>(eval);
-}
-
-
-void Nnue::set_board(const chess::Board& board) {
-    idx_ = 0;
-    for (const auto perspective : {chess::Color::WHITE, chess::Color::BLACK}) {
-        const bool mirror = needs_mirroring(board.king_square(perspective));
-        const auto bucket = king_bucket(board.king_square(perspective), perspective);
-
-        finny_table[perspective][mirror][bucket].sync(
-            params->W0[bucket], board, perspective, mirror
-        );
-        accumulators[idx_].refresh_from(finny_table[perspective][mirror][bucket], perspective);
-    }
-}
-
-void Nnue::make_move(const chess::Board& board, chess::Move move) {
-    assert(idx_ < MAX_DEPTH - 1);
-    idx_++;
-
-    const auto stm = board.stm();
-    const auto from_sq = move.from();
-    const auto to_sq = move.to();
-    const auto from_piece = board.at(from_sq);
-    const auto to_piece = board.at(to_sq);
-    auto new_king_sq = move.to();  // assuming from_piece == KING
-    assert(from_piece != chess::Piece::NONE);
-
-    accumulators[idx_].prepare_updates();
-
-    // remove moving piece
-    accumulators[idx_].rem_piece(from_piece, from_sq);
-
-    // add moved/promoted piece
-    if (move.type() == chess::Move::PROMOTION) {
-        const auto promo = chess::Piece(move.promotion_type(), stm);
-        accumulators[idx_].add_piece(promo, to_sq);
-    } else if (move.type() == chess::Move::CASTLING) {
-        assert(from_piece.type() == chess::PieceType::KING);
-        assert(to_piece.type() == chess::PieceType::ROOK);
-
-        const bool is_king_side = to_sq > from_sq;
-        new_king_sq = chess::Square::castling_king_dest(is_king_side, stm);
-        const auto rook_sq = chess::Square::castling_rook_dest(is_king_side, stm);
-        accumulators[idx_].add_piece(from_piece, new_king_sq);
-        accumulators[idx_].add_piece(to_piece, rook_sq);
-    } else
-        accumulators[idx_].add_piece(from_piece, to_sq);
-
-    // add captured piece/ep pawn/castling rook
-    if (to_piece != chess::Piece::NONE)
-        accumulators[idx_].rem_piece(to_piece, to_sq);
-    else if (move.type() == chess::Move::ENPASSANT) {
-        assert(from_piece.type() == chess::PieceType::PAWN);
-
-        const auto ep_pawn = from_piece.color_flipped();
-        const auto ep_sq = to_sq.ep_square();
-        accumulators[idx_].rem_piece(ep_pawn, ep_sq);
-    }
-
-    // need refresh if previous accumulator needs refresh or we change mirroring/bucket
-    if (accumulators[idx_ - 1].get_psq_state(stm) == NnueAccumulator::PSQState::REFRESH
-        || (from_piece.type() == chess::PieceType::KING
-            && ((needs_mirroring(from_sq) != needs_mirroring(new_king_sq))
-                || (king_bucket(from_sq, stm) != king_bucket(new_king_sq, stm)))))
-        accumulators[idx_].set_psq_state(stm, NnueAccumulator::PSQState::REFRESH);
-}
-
-void Nnue::unmake_move() {
-    assert(idx_ > 0);
-    idx_--;
-}
-
-
-bool Nnue::needs_mirroring(chess::Square king_sq) { return king_sq.file() > chess::File::D; }
-
-i32 Nnue::king_bucket(chess::Square king_sq, chess::Color perspective) {
-    const bool mirror = needs_mirroring(king_sq);
-    const auto sq = ((mirror) ? king_sq.mirrored() : king_sq).relative(perspective);
-    return BUCKETS[4 * sq.rank() + sq.file()];
-}
-
-void Nnue::lazy_update(const chess::Board& board, chess::Color perspective) {
-    // find first clean/needs_refresh accumulator
-    i32 clean_idx = idx_;
-    while (accumulators[clean_idx].get_psq_state(perspective) == NnueAccumulator::PSQState::DIRTY)
-        clean_idx--;
-
-    // horizontal mirroring and king bucket
-    const bool mirror = needs_mirroring(board.king_square(perspective));
-    const auto bucket = king_bucket(board.king_square(perspective), perspective);
-
-    // if an accumulator needs refresh, refresh at idx_ since we don't know the board at clean_idx
-    if (accumulators[clean_idx].get_psq_state(perspective) == NnueAccumulator::PSQState::REFRESH) {
-        finny_table[perspective][mirror][bucket].sync(
-            params->W0[bucket], board, perspective, mirror
-        );
-        accumulators[idx_].refresh_from(finny_table[perspective][mirror][bucket], perspective);
-        return;
-    }
-
-    // apply update up the stack
-    while (clean_idx++ < idx_)
-        accumulators[clean_idx].apply_updates(
-            accumulators[clean_idx - 1], params->W0[bucket], perspective, mirror
-        );
-}
-
-void Nnue::activate_l0(
-    const i16 acc[L1_SIZE], u8 l0_out[L1_SIZE / 2], [[maybe_unused]] SparseIterator& sp
-) const {
-    constexpr i32 n_pairs = L1_SIZE / 2;
-
-#ifdef USE_SIMD
-    constexpr i32 regw16 = ALIGNMENT / sizeof(i16);
-    static_assert(L1_SIZE % (8 * regw16) == 0);
-
-    constexpr i32 n_chunks = n_pairs / regw16;
-    const VecI16 zs = zero_i16();
-    const VecI16 qa = full_i16(QA);
-
-    for (i32 i = 0; i < n_chunks; i += 4) {
-        // compute 4 * regw16 values of the pairwise mul at once, input in [0, QA]
-        const VecI16 acc0_v0 = clamp_i16(load_i16(&acc[(i + 0) * regw16]), zs, qa);
-        const VecI16 acc1_v0 = clamp_i16(load_i16(&acc[(i + 1) * regw16]), zs, qa);
-        const VecI16 acc2_v0 = clamp_i16(load_i16(&acc[(i + 2) * regw16]), zs, qa);
-        const VecI16 acc3_v0 = clamp_i16(load_i16(&acc[(i + 3) * regw16]), zs, qa);
-        const VecI16 acc0_v1 = clamp_i16(load_i16(&acc[(i + 0) * regw16 + n_pairs]), zs, qa);
-        const VecI16 acc1_v1 = clamp_i16(load_i16(&acc[(i + 1) * regw16 + n_pairs]), zs, qa);
-        const VecI16 acc2_v1 = clamp_i16(load_i16(&acc[(i + 2) * regw16 + n_pairs]), zs, qa);
-        const VecI16 acc3_v1 = clamp_i16(load_i16(&acc[(i + 3) * regw16 + n_pairs]), zs, qa);
-
-        const VecI16 pw0 = mulhi_i16(lshift_i16(acc0_v0, 7), acc0_v1);
-        const VecI16 pw1 = mulhi_i16(lshift_i16(acc1_v0, 7), acc1_v1);
-        const VecI16 pw2 = mulhi_i16(lshift_i16(acc2_v0, 7), acc2_v1);
-        const VecI16 pw3 = mulhi_i16(lshift_i16(acc3_v0, 7), acc3_v1);
-
-        // packus will interleave every 8 values thus l1w must be permuted, output in [0, 127]
-        const VecU8 out0 = pack_u8_i16(pw0, pw1);
-        const VecU8 out1 = pack_u8_i16(pw2, pw3);
-        store_u8(&l0_out[(i + 0) * regw16], out0);
-        store_u8(&l0_out[(i + 2) * regw16], out1);
-
-        // track nonzero blocks
-        sp.add_nonzeros(out0, out1);
-    }
-#else
-    for (i32 i = 0; i < n_pairs; i++) {
-        const i32 acc_v0 = min(max(acc[i], i16(0)), i16(QA));
-        const i32 acc_v1 = min(max(acc[i + n_pairs], i16(0)), i16(QA));
-
-        // simulate mulhi, assuming non-permuted weights
-        l0_out[i] = ((acc_v0 << 7) * acc_v1) >> 16;
-    }
-#endif
-
-#ifdef MEASURE_SPARSITY
-    record_ft_activations(l0_out);
-#endif
-}
-
-void Nnue::forward_l1(
-    const u8 l0_out[L1_SIZE],
-    i32 l1_out[L2_SIZE],
-    [[maybe_unused]] const SparseIterator& sp,
-    i32 bucket_idx
-) const {
-#ifdef USE_SIMD
-    // get nnz
-    constexpr i32 regw8 = ALIGNMENT / sizeof(i8);
-    constexpr i32 regw32 = ALIGNMENT / sizeof(i32);
-    static_assert(L2_SIZE % regw32 == 0);
-    static_assert(L1_SIZE % 16 == 0);
-
-    const i32 nnz = sp.count();
-    const i32 nnz4 = (nnz / 4) * 4;
-
-    // compute l1 matmul
-    constexpr i32 n_chunks = L2_SIZE / regw32;
-    VecI32 l1_pre[n_chunks][4];
-
-    #pragma GCC unroll 32
-    for (i32 r = 0; r < n_chunks; r++) {
-        l1_pre[r][0] = zero_i32();
-        l1_pre[r][1] = zero_i32();
-        l1_pre[r][2] = zero_i32();
-        l1_pre[r][3] = zero_i32();
-    }
-
-    for (i32 nnz_id = 0; nnz_id < nnz4; nnz_id += 4) {
-        const i32 tile_id0 = sp.index(nnz_id + 0);
-        const i32 tile_id1 = sp.index(nnz_id + 1);
-        const i32 tile_id2 = sp.index(nnz_id + 2);
-        const i32 tile_id3 = sp.index(nnz_id + 3);
-
-        const VecU8 inputs0 = tile_u8(&l0_out[4 * tile_id0]);
-        const VecU8 inputs1 = tile_u8(&l0_out[4 * tile_id1]);
-        const VecU8 inputs2 = tile_u8(&l0_out[4 * tile_id2]);
-        const VecU8 inputs3 = tile_u8(&l0_out[4 * tile_id3]);
-
-        for (i32 r = 0; r < n_chunks; r++) {
-            const VecI8 weights0 = load_i8(&params->W1[bucket_idx][tile_id0][r * regw8]);
-            const VecI8 weights1 = load_i8(&params->W1[bucket_idx][tile_id1][r * regw8]);
-            const VecI8 weights2 = load_i8(&params->W1[bucket_idx][tile_id2][r * regw8]);
-            const VecI8 weights3 = load_i8(&params->W1[bucket_idx][tile_id3][r * regw8]);
-
-            l1_pre[r][0] = dpbusd_i32(l1_pre[r][0], inputs0, weights0);
-            l1_pre[r][1] = dpbusd_i32(l1_pre[r][1], inputs1, weights1);
-            l1_pre[r][2] = dpbusd_i32(l1_pre[r][2], inputs2, weights2);
-            l1_pre[r][3] = dpbusd_i32(l1_pre[r][3], inputs3, weights3);
-        }
-    }
-
-    for (i32 nnz_id = nnz4; nnz_id < nnz; nnz_id++) {
-        const i32 tile_id = sp.index(nnz_id);
-        const VecU8 inputs = tile_u8(&l0_out[4 * tile_id]);
-
-        for (i32 r = 0; r < n_chunks; r++) {
-            const VecI8 weights = load_i8(&params->W1[bucket_idx][tile_id][r * regw8]);
-            l1_pre[r][0] = dpbusd_i32(l1_pre[r][0], inputs, weights);
-        }
-    }
-
-    // activate l1
-    const VecI32 zs = zero_i32();
-    const VecI32 qs = full_i32(QC << L1_SHIFT);
-
-    for (i32 r = 0; r < n_chunks; r++) {
-        const VecI32 pre0 = add_i32(l1_pre[r][0], l1_pre[r][1]);
-        const VecI32 pre1 = add_i32(l1_pre[r][2], l1_pre[r][3]);
-
-        // apply screlu and downshift into QC^2 space
-        const VecI32 bias = load_i32(&params->b1[bucket_idx][r * regw32]);
-        const VecI32 pre = add_i32(add_i32(pre0, pre1), bias);
-        const VecI32 crelu = clamp_i32(pre, zs, qs);
-        const VecI32 screlu = rshift_i32(mullo_i32(crelu, crelu), 2 * L1_SHIFT);
-        store_i32(&l1_out[r * regw32], screlu);
-    }
-#else
-    constexpr i32 n_tiles = L1_SIZE / 4;
-    i32 l1_pre[L2_SIZE] = {0};
-
-    // compute l1 matmul
-    for (i32 i = 0; i < n_tiles; i++)
-        for (i32 j = 0; j < L2_SIZE; j++)
-            for (i32 k = 0; k < 4; k++)
-                l1_pre[j] += l0_out[4 * i + k] * params->W1[bucket_idx][i][4 * j + k];
-
-    // activate l1
-    for (i32 i = 0; i < L2_SIZE; i++) {
-        const i32 bias = params->b1[bucket_idx][i];
-        const i32 pre = l1_pre[i] + bias;
-        const i32 crelu = min(max(pre, 0), QC << L1_SHIFT);
-        const i32 screlu = (crelu * crelu) >> (2 * L1_SHIFT);
-        l1_out[i] = screlu;
-    }
-#endif
-}
-
-void Nnue::forward_l2l3(const i32 l1_out[L2_SIZE], i64& l3_out, i32 bucket_idx) const {
-#ifdef USE_SIMD
-    // compute l2 matmul
-    constexpr i32 regw32 = ALIGNMENT / sizeof(i32);
-    static_assert(L3_SIZE % regw32 == 0);
-
-    constexpr i32 n_chunks = L3_SIZE / regw32;
-    VecI32 l2_pre[n_chunks];
-
-    #pragma GCC unroll 32
-    for (i32 r = 0; r < n_chunks; r++) l2_pre[r] = load_i32(&params->b2[bucket_idx][r * regw32]);
-
-    for (i32 i = 0; i < L2_SIZE; i++) {
-        // l2_pre += W2[:, i] * l1_out[i], inputs in QC^2 space, weights in QC space
-        VecI32 input = full_i32(l1_out[i]);
-        VecI32 weights[n_chunks];
-
-        #pragma GCC unroll 32
-        for (i32 r = 0; r < n_chunks; r++)
-            weights[r] = load_i32(&params->W2[bucket_idx][i][r * regw32]);
-
-        #pragma GCC unroll 32
-        for (i32 r = 0; r < n_chunks; r++) l2_pre[r] = fmadd_i32(input, weights[r], l2_pre[r]);
-    }
-
-    // activate l2 and compute l3 matmul
-    const VecI32 zs = zero_i32();
-    const VecI32 qs = full_i32(QC * QC * QC);
-
-    l3_out = params->b3[bucket_idx];
-    VecI32 sums = zero_i32();
-
-    for (i32 r = 0; r < n_chunks; r++) {
-        const VecI32 crelu = clamp_i32(l2_pre[r], zs, qs);
-        const VecI32 weights = load_i32(&params->W3[bucket_idx][r * regw32]);
-        sums = fmadd_i32(crelu, weights, sums);
-    }
-
-    l3_out += hadd_i32(sums);
-#else
-    i32 l2_out[L3_SIZE];
-    for (i32 i = 0; i < L3_SIZE; i++) l2_out[i] = params->b2[bucket_idx][i];
-
-    // compute l2 matmul
-    for (i32 i = 0; i < L2_SIZE; i++)
-        for (i32 j = 0; j < L3_SIZE; j++) l2_out[j] += l1_out[i] * params->W2[bucket_idx][i][j];
-
-    // activate l2 and compute l3 dotprod
-    l3_out = params->b3[bucket_idx];
-    for (i32 i = 0; i < L3_SIZE; i++) {
-        const i32 crelu = min(max(l2_out[i], 0), QC * QC * QC);
-        l3_out += crelu * params->W3[bucket_idx][i];
-    }
-#endif
-}
-
-
-#ifdef MEASURE_SPARSITY
-u64 Nnue::save_ft_activations() {
-    ofstream outfile("ft_activations.json");
-    if (!outfile.is_open()) throw runtime_error("could not save ft activations");
-
-    outfile << "[ ";
-    for (i32 i = 0; i < L1_SIZE / 2; i++) {
-        if (i != 0) outfile << ", ";
-        outfile << ft_activations[i];
-    }
-    outfile << " ]\n";
-    outfile.close();
-
-    // called once for each perspective
-    assert(total_calls % 2 == 0);
-    return total_nnz / (total_calls / 2);
-}
-
-void Nnue::record_ft_activations(const u8 l0_out[L1_SIZE / 2]) {
-    // track activations
-    for (i32 i = 0; i < L1_SIZE / 2; i++)
-        if (l0_out[i] != 0) ft_activations[i]++;
-
-    // track nonzero blocks
-    for (i32 i = 0; i < L1_SIZE / 2; i += 4) {
-        bool nonzero = false;
-        for (i32 j = 0; j < 4; j++) {
-            if (l0_out[i + j] != 0) {
-                nonzero = true;
-                break;
-            }
-        }
-        if (nonzero) total_nnz++;
-    }
-    total_calls++;
-}
-#endif
